APIze docker-companion functions
